Add LemonSqueezy.js SDK and remove manual HTTP calls
Replaced manual HTTP API calls to the LemonSqueezy API with the LemonSqueezy.js SDK. This means that we have replaced a lengthy, error-prone, and manually handled API call process with a refined SDK. This not only increases maintainability, code readability, and error handling but also uses SDK built-in functions to achieve the same functionality. For instance, in 'src/lib/ls/get-subscription.ts', replaced method-based fetch calls with SDK's getSubscription function. Removed Stripe payment gateway package as we have migrated to a new LemonSqueezy gateway.

Finally, refactored some parts of the business logic to comply with SDK usage, such as passing subscriptionId as a number instead of a string.