Update Next.js dependencies to version 14.0.0 (#100)
* Update Next.js dependencies to version 14.0.0

Updated all Next.js-related dependencies to version 14.0.0 in the package-lock.json file. The update was required to take advantage of the new features, improvements, and bug fixes provided by the new version. Also updated the Stripe-js dependency to version 2.1.10 for better stability and improved performance.

* Upgrade Node version to 18 in GitHub workflow

This commit upgrades the Node.js version used in our GitHub Actions workflows from version 16 to 18.
This upgrade was necessary to align with the new Node.js version in our project's stack, enabling us to take advantage of the new features and improvements included in the latest release, and to ensure compatibility with the updated Next.js dependencies.

* Added Email OTP sign in method