Handle case of missing inviter displayName
Updated logic in invite-members to handle situations when inviter does not have a displayName. Now, we fallback to using their email. Before this change, if inviter did not have a displayName, it resulted in null value, causing errors in further processing. Now we explicitly assign an empty string in absence of a displayName.