0.7.14 (#105)
* 0.7.14

Update package versions and add organization metadata

Package-lock.json and package.json have been updated to secure the latest package versions for improved functionality and security fixes. Specifically, the next-supabase-saas-kit version has been raised to 0.7.13 and significant package updates include "edge-csrf" to ^1.0.6, "recharts" to ^2.9.3, "@types/eslint" to ^8.44.7, and "@types/i18next-fs-backend" to ^1.1.5.

Additionally, organizationUid metadata has been added to the Stripe checkout creation process to facilitate better management of subscriptions at the organizational level.

Lastly, a new metadataBase path, derived from the site URL, has been added to the application layout for more comprehensive site metadata.

* Add user and organization delete functionality in admin

This commit implements the functionality for administrators to delete a user or an organization. The change was needed to allow administrators more control over the user and organization database. Functionality for deleting a user or an organization and its data has been added with safeguards, such as verifying the request comes from an administrator, and notifying the administrator that the action is irreversible. The delete functionality is available in the drop-down menu for every user and organization in the administrator view.

* Remove getRedirectUrl and update returnUrl to next

Deleted the getRedirectUrl function as it was unnecessary, and updated 'returnUrl' to 'next' in OAuthProviders.tsx to enhance readability and understanding of code. The function was deemed redundant and its removal will improve efficiency while clarity was improved by updating variable names.

* Refactor and improve BanUserModal and PostList components

The PostsList component was deleted because it was obsolete, and the logic was directly implemented in the BlogPage. This decision was taken for better code readability and to avoid unnecessary abstraction.

The BanUserModal component was significantly modified to improve user experience and ensure edge cases are handled. An error boundary was added to improve error handling. Also, a form for banning confirmation was improved by adding a requirement of typing "BAN" to avoid accidental actions. Asserted checks for avoiding destructive actions on the admin's account were also added in actions.server file.

In addition, the user's email is used as the display text in BanUserModal, if it exists; Otherwise, the user's phone is shown. Pagination in the admin's UsersPage was fixed by correctly parsing the result of the organizations query. The blog page only shows posts which are live. In the development environment, it shows all posts.

* Update URL for organization members view

The URL in the admin users page leading to the respective organization was adjusted to lead directly to the members page of the organization. This change provides more convenience for admins, as it directly leads them to the member-specific view rather than the organization's general overview.