0.7.15 (#106)
Update Sentry, Cypress, rehype-autolink-headings, and prettier-plugin-tailwindcss

Sentry is updated to version "^7.79.0" across multiple packages for bug fixes and new feature introductions provided by the newer versions.
Cypress is updated to "^13.5.0", and rehype-autolink-headings to "^7.1.0" for the same reasons.
Lastly, prettier-plugin-tailwindcss is updated to "^0.5.7" for improved code formatting.
(Module versions were updated in package-lock.json.)