0.9.2 (#115)
* 0.9.2

Update package dependencies

Updated the versions of several package dependencies to match or exceed the required versions for smooth operation of the project. The new versions include bug fixes and new features, which enhance the stability, reliability, and performance of the project.

Changes include:
- The version of "next-supabase-saas-kit" has been updated from 0.9.1 to 0.9.2 for both the project and its dependencies.
- "@next/bundle-analyzer" version updated from 14.0.2 to 14.0.3
- The version of "@next/bundle-analyzer" dependency in "next" package has been updated from 14.0

* Upgrade several dependencies to their latest versions

Updated several packages to their latest versions in package-lock.json. These updates include sentry, stripe, supabase, react, and other packages. The changes were performed to benefit from the latest features, bug fixes, and improvements provided by these newer versions. It's part of routine maintenance and package upgrades for the project.

* Update Sentry packages and sonner package

Update versions of `@sentry/node`, `@sentry/react`, `@sentry/tracing` and `sonner` packages to increase performance and fix existing bugs. A minor upgrade in the `@types/node` package was also done. The changes are following semver so no breaking changes are expected.

* Fix RichColors prop in Toaster and add SideDialog

Mistake in passing richColors prop value to Sonner component in Toaster.tsx file was fixed. A new SideDialog component was added to ui in the core directory. The SideDialog component utilises 'react-dialog' of '@radix-ui'. It is optionally modal, with autofocus prevented, and carries custom styling. The new dialog slides in from right with animation.